backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
python-dotenv>=1.0
celery>=5.3
redis>=4.5
httpx>=0.24
django-cors-headers>=4.0
pytest
pytest-django
django-celery-beat>=2.6.0
drf-spectacular
drf-spectacular[sidecar]
Faker==23.3.0